force nvq++ to be verbose.
Add more verbosity

Use --gcc-install-dir option instead of --gcc-toolchain. The directory
is there, but the tool is ignoring it.

Try to make the triples for clang and gcc match by hackery.

Undate driver to support -Xcudaq.

Remove 32 directory which has broken links from the installer.

Try rm + ls on the directory.

Remove quotes