[NFC] Move copious inline code from out of the BaseRemoteRESTQPU.h file.
The entire BaseRemoteRESTQPU class was written inline in the header file
in the common directory. This change moves the larger member functions
out of the header file and into a .cpp implementation file.

The implementation is placed with where it was used in RemoteRESTQPU.cpp
to avoid circular refernces in the linker step.