[core] Separate compilation, supporting C++ host side function references.
These changes add a new wrapper class to support interfacing between
kernels. Kernels on the device side have their own names and calling
conventions. Meanwhile, the C++ host code can capture references to
these kernels. The C++ compiler assumes it can erase the functions
completely and degenerate any reference to them as a pointer.
Furthermore, it may inline or wrap this code in a thicket of template
instantiations. But in order to be useful on the device side, the CUDA-Q
runtime must be able to determine which kernel was wrapped in C++ code,
lookup the device side code, and "link-time" optimize these calls, doing
calling convention conversions, etc. Furthermore, what must be done is
flavored by the execution environment.