Adds crude support for apply_noise with broadcast.
There are still issues here with the headers checking the number of
qubits allowed.

The biggest issue is that we need hackery to rebuild a qvector data
type from the device side. qvector itself is pretty constant. One can't
make a copy of a qvector and the qubit identities are preset, which
means we can't easily copy these ids out of the QIR representation.
So this PR contains several shadow types to step around the C++ compiler
checks that would otherwise block the use of qvector values.

Signed-off-by: Eric Schweitz <eschweitz@nvidia.com>